Set time zone from HK metadata in FHIR resources (#50)
# Set time zone from HK metadata in FHIR resources

## :recycle: Current situation & Problem
The HealthKitOnFHIR library currently assumes that HealthKit samples
were created in the current time zone of the device at the time of
mapping and applies the corresponding UTC offset. While this is true for
most situations, it is possible that older HealthKit data may be mapped
using the library that was created while the user was in a different
time zone.

For example, if a sample was originally created by HealthKit in San
Francisco on December 1st, 2024 at 9:00am (-08:00), but then mapped by
HealthKitOnFHIR later on in New York, the timestamp would read
2024-12-01T12:00:00-05:00. While technically these are the same point in
time, they are offset differently.

The information regarding the original timezone of the sample may be
stored by HealthKit in the object's metadata via
[HKMetadataKeyTimeZone](https://developer.apple.com/documentation/healthkit/hkmetadatakeytimezone),
if this was written by the original application or device that created
the data, and we should use it to apply the appropriate offset.


## :gear: Release Notes 
- If a time zone has been defined in
[HKMetadataKeyTimeZone](https://developer.apple.com/documentation/healthkit/hkmetadatakeytimezone),
it is set when serializing HealthKit samples.

## :white_check_mark: Testing
Unit tests have been updated to test that the appropriate offset is
applied for multiple time zones specified via metadata and test that the
offset of the current time zone is applied if none is specified.

### Code of Conduct & Contributing Guidelines 

By submitting creating this pull request, you agree to follow our [Code
of
Conduct](https://github.com/StanfordBDHG/.github/blob/main/CODE_OF_CONDUCT.md)
and [Contributing
Guidelines](https://github.com/StanfordBDHG/.github/blob/main/CONTRIBUTING.md):
- [X] I agree to follow the [Code of
Conduct](https://github.com/StanfordBDHG/.github/blob/main/CODE_OF_CONDUCT.md)
and [Contributing
Guidelines](https://github.com/StanfordBDHG/.github/blob/main/CONTRIBUTING.md).